Sensor Server is Operational
The sensor server now accepts requests for other servers to
receive updates when new sensor data comes in.

All screen drawing code was moved into user_prompt. This
probably needs to be seperated a bit more still. A task
in user_prompt is now a client of the sensor server that
prints recent sensors to the screen.