Reduce Number of Location Updates We Print
Running two trains on the track causes the terminal
putc queue to fill up, so now we only print a location
update if it has moved by more than 1 cm.

Also revert FINIDNG_LOCATION_SPEED to 2 becuase at 1 the
train gets stuck on the sensor and we can't add a second train.