Stop Ignoring Sensor Updates when Stopping
So this caused issues when the train gets stuck. Since
we can travel quite far while stopping (i.e. on reverses) the
location becomes very inaccurate when the trian gets stuck while
reversing.

Instead, I just stopped resetting um_past_node when we hit a
sensor if we are stopping. I measured the results and they
were just as good.