BUG FIX: Missing Case in Edge Groups
In the reservation server we associate a "group" of
edges for each edge that we reserve/free at the same
time.

There was a missing case where an edge could be part of
two branch nodes.

Also added names to track_edge_arrays so that we know
which one failed when an assert fails.